% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_custom_effect_sizes.R
\name{add_custom_effect_sizes}
\alias{add_custom_effect_sizes}
\title{Calculate effect sizes}
\usage{
add_custom_effect_sizes(
  formula,
  name,
  derivative,
  calculation_type,
  vars_required,
  label = NA,
  filtering_rules = NA,
  special_warnings = NA,
  SAFE_family,
  add_to_stable_formulas = TRUE
)
}
\arguments{
\item{formula}{Vector of formulas as character strings. If mathematical functions are used (e.g., xxxxx()) they must be namespaced.}

\item{name}{Name of effect size}

\item{derivative}{"first", "second" order derivative.}

\item{calculation_type}{Character vector of equal length as 'formula' specifying whether formula is for 'effect_size' or 'sampling_variance'. Required}

\item{vars_required}{xxxx Required. Character string of required input variables. E.g., x1, x2, n1, n2, etc.}

\item{label}{xxxx Optional character string to label results}

\item{filtering_rules}{xxxxx (NA or a character string that could be executed as R code. See details)}

\item{special_warnings}{xxxx (optional)}

\item{SAFE_family}{Options include: "1_normal", "2_multivariate_normal", "4_multivariate_normal", "4_multivariate_normal_wishart", "2_binomial", "2_multinomial"}

\item{add_to_stable_formulas}{Add to already modified formulas or to user-defined formulas?}
}
\value{
A data.table with effect sizes and sample variances
}
\description{
Blah blah blah
}
