% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{add_labels}
\alias{add_labels}
\title{Apply labels to multiple variables on a data frame}
\usage{
add_labels(data, ...)
}
\arguments{
\item{data}{A data.frame or tibble}

\item{...}{Named parameters in the form of variable = 'label'}
}
\value{
data with variable labels applied
}
\description{
This function allows a user to apply several labels to a dataframe at once.
}
\examples{
add_labels(
    mtcars,
    mpg = "Miles Per Gallon",
    cyl = "Cylinders"
  )

}
