% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codelists.R
\name{apply_codelist}
\alias{apply_codelist}
\title{Apply codelists}
\usage{
apply_codelist(dataset, metacore_slice)
}
\arguments{
\item{dataset}{dataset to apply the codelist to}

\item{metacore_slice}{filtered metacore object to only have the dataset you are looking at}
}
\value{

}
\description{
Apply codelists
}
