% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_effect.R
\name{calc_effect}
\alias{calc_effect}
\title{Evaluate formulas}
\usage{
calc_effect(formulas, input)
}
\arguments{
\item{formulas}{Formula to evalulate}

\item{input}{List of named input variables to evaluate formula on}
}
\value{
A data.table with effect sizes and sample variances
}
\description{
Used internally
}
