% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{CalculateAUC}
\alias{CalculateAUC}
\title{CalculateAUC}
\usage{
CalculateAUC(x, y)
}
\arguments{
\item{x}{a list of numbers}

\item{y}{a list of numbers matching the number of elements of x}
}
\description{
Calculate area under the curve for function y(x)
}
\examples{
x <- c(1,3,7,14)
y <- c(6.3,5.6,4.5,2.6)
ysd <- c(7,8,5,3)

CalculateAUC(y,x)

}
