% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{CalculateAUCSD}
\alias{CalculateAUCSD}
\title{CalculateAUCSD}
\usage{
CalculateAUCSD(y, x, ysd = NULL)
}
\arguments{
\item{ysd}{}
}
\description{
Calculate AUC SD
}
\examples{
x <- c(1,3,7,14)
y <- c(6.3,5.6,4.5,2.6)
ysd <- c(7,8,5,3)

CalculateAUC(y,x,ysd)

}
