% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supp.R
\name{combine_supp}
\alias{combine_supp}
\title{Combine the Domain and Supplemental Qualifier}
\usage{
combine_supp(dataset, supp)
}
\arguments{
\item{dataset}{Domain dataset}

\item{supp}{Supplemental Qualifier dataset}
}
\value{
a dataset with the supp variables added to it
}
\description{
Combine the Domain and Supplemental Qualifier
}
\examples{
library(safetyData)
library(tibble)
combine_supp(sdtm_ae, sdtm_suppae)  \%>\% as_tibble()
}
