% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_formulas.R
\docType{data}
\name{conversion_formulas}
\alias{conversion_formulas}
\title{Effect size & sampling variance formulas}
\format{
A data.table containing the conversion formulas used by \code{convert_effect_sizes}
\describe{
\item{from_effect}{XXXX}
\item{to_effect}{XXXX}
\item{vars_required}{XXXXX}
\item{message}{XXXXX}
\item{equations_used}{XXXXXXX}
\item{calc_type}{XXXXXX}
\item{formula}{XXXXXXX}
\item{exec_formula}{XXXXXX}
\item{label}{XXXXXX}
...
}
}
\source{
XXXXXXX
}
\usage{
data(conversion_formulas)
}
\description{
Formulas blah blah blah to calculate blah blah blah
}
\keyword{datasets}
