% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_size.R
\name{eff_size}
\alias{eff_size}
\title{Calculate effect sizes}
\usage{
eff_size(
  ...,
  effect_type = NULL,
  data = NULL,
  bind = TRUE,
  paired = FALSE,
  default_formulas = TRUE,
  formula_path = "data/effect_size_formulas.csv",
  SAFE = FALSE,
  SAFE_boots = 1e+06,
  SAFE_max_secs = 15,
  n_cores = 1,
  SAFE_distribution = NULL,
  custom_sigma_matrix = NULL,
  verbose = T
)
}
\arguments{
\item{...}{variables required by effect size calculation (e.g., x, x1, x2, sd1, sd2, see Description)}

\item{effect_type}{Single length character specifying effect size. Currently supported effect sizes include: SMD, SMDH, lnRoM, lnCVR, lnVR, lnM, Zr, lnHWE}

\item{data}{An optional data frame containing the variables necessary for calculating effect size.}

\item{bind}{Should the function bind the effect sizes to the inputted data frame? Logical}

\item{paired}{Use paired version of effect size? If so, then inputs will differ from unpaired (e.g., \code{n} instead of \code{n1} and \code{n2}).}

\item{default_formulas}{Logical. Whether to return default estimators. If FALSE (recommended and the function's default), will return first and second derivative estimators. Note that second refers to bias-corrected estimates (which are not always less biased). See Details.}

\item{SAFE}{Logical. Whether to conduct SAFE bootstrapping of effect size. This is essential to calculate lnM in boundary scenarios}

\item{SAFE_boots}{Numeric. Number of bootstraps. Default is 1e6.}

\item{SAFE_max_secs}{Numeric. Number of seconds to conduct SAFE bootstrapping before timing out. See Details.}

\item{n_cores}{Numeric. Number of cores for parallel processing of SAFE calculation. See Details.}

\item{SAFE_distribution}{Character. Distribution to use for SAFE bootstrapping. If unspecified, will use default distribution. See Details.}

\item{verbose}{How chatty do you want the function to be? Logical}

\item{sigma_matrix}{Matrix. Optional custom sigma_matrix for SAFE bootstrapping (experimental)}
}
\value{
A data.table with effect sizes and sample variances.
}
\description{
This function calculates effect sizes for meta-analysis.
}
\details{
It currently supports XXXX effect size point estimates and their sampling variance.
Point estimates are denoted by \code{yi} and variance estimates by \code{vi}. If \code{default_formulas = FALSE} this function will return multiple estimators
of the same effect size. \verb{_first} indicate the definition formula of the effect size. For example, for SMD (standardized mean difference),
this would be Cohen's d. \verb{_second} denote second-order derivatives that account for small-sample bias---but actually may be more biased.
For example, for SMD, the \code{vi_second} is more biased than \code{vi_first}. Likewise, different estimators may have different assumptions.

If \code{default_formulas = TRUE} then the function returns the "best" estimators, following the same logic as \code{metafor::escalc} (Viechtbauer 2010).

These are described at length in Lundgren et al. XXXXX
}
\references{
W. Viechtbauer, Conducting meta-analyses in R with the metafor package. Journal of Statistical Software. (2010). https://doi.org/10.18637/jss.v036.i03.

\strong{ADD a message in function regarding which estimator is preferred if default_formulas = FALSE?}
}
