% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_formulas.R
\docType{data}
\name{effect_formulas}
\alias{effect_formulas}
\title{Effect size & sampling variance formulas}
\format{
A data.table with X rows and Y columns
\describe{
\item{effect_size}{XXXX}
\item{derivative}{XXXXX}
\item{SAFE_family}{XXXXX}
\item{vars_required}{XXXXX}
\item{cloud_filtering_rules}{XXXXX}
\item{special_warnings}{XXXXX}
\item{default_safe_family}{XXXXX}
\item{calc_type}{XXXXX}
\item{formula}{XXXXX}
\item{label}{XXXXX}
\item{exec_formula}{XXXXX}
\item{paired_design}{XXXXX}
\item{reference}{XXXXX}
\item{escalc_name}{XXXX}
\item{must_use_SAFE}{XXXXX}
\item{equations_used}{XXXXX}
\item{n_versions}{XXXXX}
\item{default}{XXXXX}
\item{SAFE_bias_correction}{XXXXX}
...
}
}
\source{
XXXXXXX
}
\usage{
data(effect_formulas)
}
\description{
Formulas blah blah blah to calculate blah blah blah
}
\keyword{datasets}
