% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{get_variables}
\alias{get_variables}
\title{Internal functions to get variables from a dataset list}
\usage{
get_variables(x, ds_list, keep)
}
\arguments{
\item{x}{Dataset with the old and new variable name and dataset name}

\item{ds_list}{List of datasets}

\item{keep}{boolean if old columns should be kept}
}
\value{
datasets
}
\description{
This function is used with `build_from_derived` to build a dataset of columns
that are pulled directly from other datasets
}
