% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metatools-package.R
\docType{package}
\name{metatools-package}
\alias{metatools}
\alias{metatools-package}
\title{metatools: Enable the Use of 'metacore' to Help Create and Check Dataset}
\description{
Uses the metadata information stored in 'metacore' objects to check and build metadata associated columns.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pharmaverse/metatools}
  \item \url{https://pharmaverse.github.io/metatools/}
  \item Report bugs at \url{https://github.com/pharmaverse/metatools/issues}
}

}
\author{
\strong{Maintainer}: Liam Hobby \email{liam.f.hobby@gsk.com}

Authors:
\itemize{
  \item Liam Hobby \email{liam.f.hobby@gsk.com}
  \item Christina Fillmore \email{christina.e.fillmore@gsk.com} (\href{https://orcid.org/0000-0003-0595-2302}{ORCID})
  \item Bill Denney
  \item Mike Stackhouse \email{mike.stackhouse@atorusresearch.com} (\href{https://orcid.org/0000-0001-6030-723X}{ORCID})
  \item Jana Stoilova \email{jana.stoilova@roche.com}
  \item Tamara Senior \email{tamara.senior@roche.com}
}

Other contributors:
\itemize{
  \item GlaxoSmithKline LLC [copyright holder, funder]
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
  \item Atorus Research LLC [copyright holder, funder]
}

}
\concept{internal}
\keyword{internal}
