% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort.R
\name{order_sort}
\alias{order_sort}
\title{Sort Columns by Order}
\usage{
order_sort(data, metacore, dataset_name = NULL)
}
\arguments{
\item{data}{dataset to sort}

\item{metacore}{metacore object that contains the specifications for the
dataset of interest.}

\item{dataset_name}{Optional string to specify the dataset. This is only
needed if the metacore object provided hasn't already been subsetted.}
}
\value{
dataset with ordered columns
}
\description{
This function sorts the dataset according to the order found in the
metacore object.
}
\examples{
 spec <- define_to_MetaCore(metacore_example("ADaM_define.xml")) \%>\%
    select_dataset("ADSL")
data <- haven::read_xpt(pkg_example("adsl.xpt"))
order_sort(data, spec)
}
