% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_cloud.R
\name{parameter_cloud}
\alias{parameter_cloud}
\title{Calculate effect sizes}
\usage{
parameter_cloud(
  sim_family,
  paired = "no",
  input,
  cloud_filtering_rule = NA,
  sigma_matrix = NULL,
  SAFE_boots = 1e+06
)
}
\arguments{
\item{sim_family}{blah}

\item{paired}{blah}

\item{input}{blah}

\item{cloud_filtering_rule}{blah}

\item{sigma_matrix}{blah}

\item{SAFE_boots}{blah}
}
\value{
A data.table with effect sizes and sample variances
}
\description{
Used internally
}
