% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pkg_example}
\alias{pkg_example}
\title{Get path to pkg example}
\usage{
pkg_example(file = NULL)
}
\arguments{
\item{file}{Name of file. If `NULL`, the example files will be listed.}
}
\description{
pkg comes bundled with a number of sample files in its `inst/extdata`
directory. This function make them easy to access
}
\examples{
pkg_example()
pkg_example("dm.xpt")
}
