% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{remove_labels}
\alias{remove_labels}
\title{Remove labels to multiple variables on a data frame}
\usage{
remove_labels(data)
}
\arguments{
\item{data}{A data.frame or tibble}
}
\value{
data with variable labels applied
}
\description{
This function allows a user to removes all labels to a dataframe at once.
}
\examples{
library(haven)
data <- read_xpt(metatools_example("adsl.xpt"))
remove_labels(data)

}
