% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAFE_calc.R
\name{SAFE_calc}
\alias{SAFE_calc}
\title{Calculate effect sizes}
\usage{
SAFE_calc(
  formulas,
  input_k,
  plugin_effect_k,
  sigma_matrix_k,
  SAFE_boots = 1e+06
)
}
\arguments{
\item{formulas}{blah}

\item{input_k}{blah}

\item{plugin_effect_k}{blah}

\item{sigma_matrix_k}{blah}

\item{SAFE_boots}{blah}
}
\value{
A data.table with effect sizes and sample variances
}
\description{
Used internally to calculate SAFE effect sizes. This function calls parameter_cloud and then transforms hyperparameters
}
