% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{variable_check}
\alias{variable_check}
\title{Check Variable Names}
\usage{
variable_check(data, metacore, dataset_name = NULL)
}
\arguments{
\item{data}{dataset to check}

\item{metacore}{metacore object that contains the specifications for the
dataset of interest.}

\item{dataset_name}{Optional string to specify the dataset. This is only
needed if the metacore object provided hasn't already been subsetted.}
}
\value{
message if the dataset matches the specification, and error otherwise
}
\description{
This function checks the variables in the dataset against the variables
defined in the metacore specifications. If everything matches the function
will return `TRUE` and a message starting everything is as expected. If there
are additional or missing variables and error will explain the discrepancies
}
\examples{
 spec <- define_to_MetaCore(metacore_example("ADaM_define.xml")) \%>\%
    select_dataset("ADSL")
data <- haven::read_xpt(pkg_example("adsl.xpt"))
variable_check(data, spec)
}
