% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_from_derived}
\alias{build_from_derived}
\title{Build a dataset from derived}
\usage{
build_from_derived(
  metacore,
  ds_list,
  dataset_name = NULL,
  predecessor_only = FALSE,
  keep = TRUE
)
}
\arguments{
\item{metacore}{Metacore object that contains the specifications for the
dataset of interest.}

\item{ds_list}{Named list of datasets that are needed to build the from}

\item{dataset_name}{Optional string to specify the dataset. This is only
needed if the metacore object provided hasn't already been subsetted.}

\item{predecessor_only}{By default `FALSE`, but if `TRUE` will only use
derivations with the origin of 'Predecessor'}

\item{keep}{Boolean to determine if the original columns should be kept. By
default `TRUE`, so if any column gets renamed between SDTM and ADAM the
original SDTM column and the renamed ADaM column are in the resulting
datasets. If `FALSE`, only the ADaM columns are kept.}
}
\value{
datset
}
\description{
This function builds a dataset out of the columns that just need to be pulled
through. So any variable that has a derivation in the format of
'dataset.variable' will be pulled through to create the new dataset. These
columns are often called 'Predecessors' in ADaM, but this is not universal so
that is optional to specify.
}
\examples{
library(metacore)
library(haven)
library(magrittr)
load(metacore_example("pilot_ADaM.rda"))
spec <- metacore \%>\% select_dataset("ADSL")
ds_list <- list(DM = read_xpt(metatools_example("dm.xpt")))
build_from_derived(spec, ds_list)
}
