% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_from_derived}
\alias{build_from_derived}
\title{Build a dataset from derived}
\usage{
build_from_derived(
  metacore,
  ds_list,
  dataset_name = NULL,
  predecessor_only = FALSE
)
}
\arguments{
\item{metacore}{metacore object that contains the specifications for the
dataset of interest.}

\item{ds_list}{Named list of datsets that are needed to build the from}

\item{dataset_name}{Optional string to specify the dataset. This is only
needed if the metacore object provided hasn't already been subsetted.}

\item{predecessor_only}{by default `FALSE`, but if `TRUE` will only use
derivations with the origin of 'Predecessor'}
}
\value{
datset
}
\description{
This function builds a dataset out of the columns that just need to be pulled
through. So any variable that has a derivation in the format of
'dataset.variable' will be pulled through to creat the new dataset. These
columns are often called 'Predecessors' in ADaM, but this is not univeral so
that is optional to specify.
}
\examples{
library(metacore)
library(haven)
library(magrittr)
metacore <- define_to_metacore(metacore_example("ADaM_define.xml"), quiet = TRUE) \%>\%
select_dataset("ADSL")
ds_list <- list(DM =  read_xpt(metatools_example("dm.xpt")))
build_from_derived(metacore, ds_list)
}
