% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supp.R
\name{build_qnam}
\alias{build_qnam}
\title{Build the observations for a single QNAM}
\usage{
build_qnam(dataset, qnam, qlabel, idvar, qeval, qorig)
}
\arguments{
\item{dataset}{Input dataset}

\item{qnam}{QNAM value}

\item{qlabel}{QLABEL value}

\item{idvar}{IDVAR variable name (provided as a string)}

\item{qeval}{QEVAL value to be populated for this QNAM}

\item{qorig}{QORIG value to be populated for this QNAM}
}
\value{
Observations structured in SUPP format
}
\description{
Build the observations for a single QNAM
}
