% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_ct_col}
\alias{check_ct_col}
\title{Check Control Terminology for a Single Column}
\usage{
check_ct_col(data, metacore, var, na_acceptable = FALSE)
}
\arguments{
\item{data}{Data to check}

\item{metacore}{A metacore object to get the codelist from. If the variable
has different codelists for different datasets the metacore object will
need to be subsetted using `select_dataset` from the metacore package.}

\item{var}{Name of variable to check}

\item{na_acceptable}{Logical value, set to `FALSE` by default, meaning
missing values are not acceptable. If set to `TRUE` then will pass check if
values are in the control terminology or are missing}
}
\value{
logical value, `TRUE` if it passes and `FALSE` if it fails
}
\description{
This function checks the column in the dataset only contains the control
terminology as defined by the metacore specification
}
\examples{
library(metacore)
library(haven)
library(magrittr)
spec <- define_to_metacore(metacore_example("ADaM_define.xml"), quiet = TRUE) \%>\%
  select_dataset("ADSL")
data <- read_xpt(metatools_example("adsl.xpt"))
check_ct_col(data, spec, TRT01PN)
check_ct_col(data, spec, "TRT01PN")
}
