% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_ct_col}
\alias{check_ct_col}
\title{Check Control Terminology for a Single Column}
\usage{
check_ct_col(data, metacore, col_name)
}
\arguments{
\item{data}{data to check}

\item{metacore}{a metacore object to get the codelist from. If the
variable has different codelists for different datasets the metacore object
will need to be subsetted using `select_dataset` from the metacore package.}

\item{col_name}{Name of column}
}
\value{
logical value, `TRUE` if it passes and `FALSE` if it fails
}
\description{
This function checks the column in the dataset only contains the control
terminology as defined by the metacore specification
}
\examples{
library(metacore)
library(haven)
spec <- define_to_MetaCore(metacore_example("ADaM_define.xml")) \%>\%
    select_dataset("ADSL")
data <- read_xpt(pkg_example("adsl.xpt"))
check_ct_col(data, spec, TRT01PN)
check_ct_col(data, spec, "TRT01PN")
}
