% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_ct_data}
\alias{check_ct_data}
\title{Check Control Terminology for a Dataset}
\usage{
check_ct_data(data, metacore)
}
\arguments{
\item{data}{dataset to check}

\item{metacore}{a subsetted metacore object. It should only have the in If the
variable has different codelists for different datasets the metacore object
will need to be subsetted using `select_dataset` from the metacore package.}
}
\value{

}
\description{
This function checks that all columns in the dataset only contains the control
terminology as defined by the metacore specification
}
