% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_ct_data}
\alias{check_ct_data}
\title{Check Control Terminology for a Dataset}
\usage{
check_ct_data(data, metacore, na_acceptable = FALSE)
}
\arguments{
\item{data}{dataset to check}

\item{metacore}{metacore object that contains the specifications for the
dataset of interest. If any variable has different codelists for different
datasets the metacore object will need to be subsetted using
`select_dataset` from the metacore package.}

\item{na_acceptable}{Logical value, set to `FALSE` by default, meaning
missing values are not acceptable. If set to `TRUE` then will pass check if
values are in the control terminology or are missing}
}
\value{
`TRUE` if all columns pass. It will error otherwise
}
\description{
This function checks that all columns in the dataset only contains the
control terminology as defined by the metacore specification
}
\examples{
library(haven)
library(metacore)
library(magrittr)
spec <- define_to_metacore(metacore_example("ADaM_define.xml"), quiet = TRUE) \%>\%
    select_dataset("ADSL")
data <- read_xpt(metatools_example("adsl.xpt"))
check_ct_data(data, spec, TRUE)
}
