% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_effect_sizes.R
\name{convert_effect_sizes}
\alias{convert_effect_sizes}
\title{Convert between effect size types}
\usage{
convert_effect_sizes(..., from = NULL, to = NULL, data = NULL, bind = FALSE)
}
\arguments{
\item{...}{variables required by effect size calculation (e.g., x, x1, x2,
sd1, sd2, see Description)}

\item{from}{Current effect size. Character of length 1.}

\item{to}{New effect size. Character of length 1.}

\item{data}{An optional data frame containing the columns}

\item{bind}{Whether to bind outputs to 'data'.}
}
\value{
A data.table with effect sizes and sample variances (if bind = TRUE). Transformed effect sizes are labeled 'yi_trans' and 'vi_trans'.
}
\description{
Function to convert effect sizes. The following conversions are available:
from lnOR -> SMD, from SMD -> lnOR, from Zr -> SMD, and from SMD -> Zr. The
conversions involving Zr can also be calculated with 'r'. The variables
required for each conversion are listed in details and will be returned by
running convert_effect_sizes()
}
