% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codelists.R
\name{create_subgrps}
\alias{create_subgrps}
\title{Create Subgroups}
\usage{
create_subgrps(ref_vec, grp_defs)
}
\arguments{
\item{ref_vec}{Vector of numeric values}

\item{grp_defs}{Vector of strings with groupings defined. Format must be
either: <00, >=00, 00-00, or  00-<00}
}
\value{
Character vector of the values in the subgroups
}
\description{
Create Subgroups
}
\examples{
create_subgrps(c(1:10), c("<2", "2-5", ">5"))
create_subgrps(c(1:10), c("<=2", ">2-5", ">5"))
create_subgrps(c(1:10), c("<2", "2-<5", ">=5"))
}
