% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAFE_calc.R
\name{.SAFE_calc}
\alias{.SAFE_calc}
\title{SAFE calculation}
\usage{
.SAFE_calc(
  formulas,
  input,
  plugin_effect,
  custom_sigma,
  SAFE_boots = 1e+06,
  SAFE_max_secs
)
}
\arguments{
\item{formulas}{blah}

\item{input}{blah}

\item{plugin_effect}{blah}

\item{custom_sigma}{blah}

\item{SAFE_boots}{blah}

\item{SAFE_max_secs}{blah}
}
\value{
A data.table with effect sizes and sample variances calculated by SAFE.
\link{I really don't like this degree of specificity of effect_type manipulation inside the function}
}
\description{
Used internally to calculate SAFE effect sizes. This function calls parameter_cloud and then transforms hyperparameters
}
