% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_size.R
\name{eff_size}
\alias{eff_size}
\title{Calculate effect sizes}
\usage{
eff_size(
  ...,
  effect_type = NULL,
  SAFE = TRUE,
  SAFE_boots = 1e+06,
  SAFE_distribution = NULL,
  sigma_matrix = NULL,
  verbose = T,
  use_custom_formulas = FALSE
)
}
\arguments{
\item{...}{variables required by effect size calculation (e.g., x, x1, x2, sd1, sd2, see Description)}

\item{effect_type}{Character of effect size. Currently supported effect sizes include SMD, lnRoM, lnCVR, OR, RR, blah blah}

\item{SAFE}{Logical. Whether to conduct SAFE bootstrapping of effect size}

\item{SAFE_boots}{Numeric. Number of bootstraps (default is 1e6)}

\item{SAFE_distribution}{Character. distribution to use for SAFE bootstrapping. See description}

\item{sigma_matrix}{Matrix. Optional custom sigma_matrix for SAFE bootstrapping.}

\item{verbose}{Verbose or not? Logical}

\item{use_custom_formulas}{Logical. Load user-customized formulas? These are saved to a local path (tools::R_user_dir(package = "effsize", which = "data")) with the add_custom_effect_sizes() function.}
}
\value{
A data.table with effect sizes and sample variances
}
\description{
Blah blah blah
}
