% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_formulas.R
\docType{data}
\name{effect_formulas}
\alias{effect_formulas}
\title{Effect size & sampling variance formulas}
\format{
A data.table with X rows and Y columns
\describe{
\item{formula}{XXXX}
\item{name}{XXXX}
\item{sim_family}{XXXXX}
...
}
}
\source{
Nakagawa et al. XXXX
}
\usage{
data(effect_formulas)
}
\description{
Formulas blah blah blah to calculate blah blah blah
}
\keyword{datasets}
