% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{get_bad_ct}
\alias{get_bad_ct}
\title{Gets vector of control terminology which should be there}
\usage{
get_bad_ct(data, metacore, var, na_acceptable = NULL)
}
\arguments{
\item{data}{Data to check}

\item{metacore}{A metacore object to get the codelist from. If the variable
has different codelists for different datasets the metacore object will
need to be subsetted using `select_dataset` from the metacore package.}

\item{var}{Name of variable to check}

\item{na_acceptable}{Logical value, set to `NULL` by default, so the
acceptability of missing values is based on if the core for the variable is
"Required" in the `metacore` object. If set to `TRUE` then will pass check
if values are in the control terminology or are missing. If set to
`FALSE`then NA will not be acceptable.e}
}
\value{
vector
}
\description{
This function checks the column in the dataset only contains the control
terminology as defined by the metacore specification. It will return all
values not found in the control terminology
}
