% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supp.R
\name{make_supp_qual}
\alias{make_supp_qual}
\title{Make Supplemntal Qualifier}
\usage{
make_supp_qual(dataset, metacore, idvar = NULL, dataset_name = NULL)
}
\arguments{
\item{dataset}{dataset the supp will be pulled from}

\item{metacore}{A subsetted metacore object to get the supp information from.
If not already subsetted then a `dataset_name` will need to be provided}

\item{idvar}{The name if the ID variable. If not provided then the `IDVAR`
and `IDVARVAL` columns of the supplemental dataset will be blank.}

\item{dataset_name}{optional name of dataset}
}
\value{
a CDISC formatted SUPP dataset
}
\description{
Make Supplemntal Qualifier
}
